+++ src/lib/components/chat/Messages/Citations.svelte
@@ -164,8 +164,8 @@
 		<button
 			class="text-xs font-medium text-gray-600 dark:text-gray-300 px-3.5 h-8 rounded-full hover:bg-gray-100 dark:hover:bg-gray-800 transition flex items-center gap-1 border border-gray-50 dark:border-gray-850/30"
 			aria-label={citations.length === 1
-				? $i18n.t('Toggle 1 source')
-				: $i18n.t('Toggle {{COUNT}} sources', { COUNT: citations.length })}
+				? $i18n.t('Toggle context usage')
+				: $i18n.t('Toggle {{COUNT}} context sources', { COUNT: citations.length })}
 			aria-expanded={showCitations}
 			on:click={() => {
 				showCitations = !showCitations;
@@ -187,9 +187,9 @@
 			{/if}
 			<div>
 				{#if citations.length === 1}
-					{$i18n.t('1 Source')}
+					{$i18n.t('Context Used')}
 				{:else}
-					{$i18n.t('{{COUNT}} Sources', {
+					{$i18n.t('{{COUNT}} Context Sources', {
 						COUNT: citations.length
 					})}
 				{/if}
@@ -204,7 +204,7 @@
 			{#each citations as citation, idx}
 				<button
 					id={`source-${id}-${idx + 1}`}
-					aria-label={$i18n.t('View source: {{name}}', {
+					aria-label={$i18n.t('View context: {{name}}', {
 						name: decodeString(citation.source.name)
 					})}
 					class="no-toggle outline-hidden flex dark:text-gray-300 bg-transparent text-gray-600 rounded-xl gap-1.5 items-center"